refactor options parsing
