+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * refactored directory structure
   * re-added JRuby and TruffleRuby
   * no C-accelleration for JRuby
+  * refactored options parsing
 
 ## 1.9.0 (2023-09-04)
   * fixed issue #139
