+++ lib/smarter_csv/smarter_csv.rb
@@ -11,11 +11,8 @@ module SmarterCSV
   class KeyMappingError < SmarterCSVException; end
 
   # first parameter: filename or input object which responds to readline method
-  def SmarterCSV.process(input, options = {}, &block) # rubocop:disable Lint/UnusedMethodArgument
-    options = default_options.merge(options)
-    options[:invalid_byte_sequence] = '' if options[:invalid_byte_sequence].nil?
-    puts "SmarterCSV OPTIONS: #{options.inspect}" if options[:verbose]
-    validate_options!(options)
+  def SmarterCSV.process(input, given_options = {}, &block) # rubocop:disable Lint/UnusedMethodArgument
+    options = process_options(given_options)
 
     headerA = []
     result = []
@@ -205,45 +202,6 @@ module SmarterCSV
     end
 
     protected
-
-    # NOTE: this is not called when "parse" methods are tested by themselves
-    def default_options
-      {
-        acceleration: true,
-        auto_row_sep_chars: 500,
-        chunk_size: nil,
-        col_sep: :auto, # was: ',',
-        comment_regexp: nil, # was: /\A#/,
-        convert_values_to_numeric: true,
-        downcase_header: true,
-        duplicate_header_suffix: nil,
-        file_encoding: 'utf-8',
-        force_simple_split: false,
-        force_utf8: false,
-        headers_in_file: true,
-        invalid_byte_sequence: '',
-        keep_original_headers: false,
-        key_mapping: nil,
-        quote_char: '"',
-        remove_empty_hashes: true,
-        remove_empty_values: true,
-        remove_unmapped_keys: false,
-        remove_values_matching: nil,
-        remove_zero_values: false,
-        required_headers: nil,
-        required_keys: nil,
-        row_sep: :auto, # was: $/,
-        silence_missing_keys: false,
-        skip_lines: nil,
-        strings_as_keys: false,
-        strip_chars_from_headers: nil,
-        strip_whitespace: true,
-        user_provided_headers: nil,
-        value_converters: nil,
-        verbose: false,
-        with_line_numbers: false,
-      }
-    end
 
     def readline_with_counts(filehandle, options)
       line = filehandle.readline(options[:row_sep])
@@ -594,22 +552,6 @@ module SmarterCSV
 
       puts "SmarterCSV found unhandled BOM! #{str.chars[0..7].inspect}"
       str
-    end
-
-    def validate_options!(options)
-      keys = options.keys
-      errors = []
-      errors << "invalid row_sep" if keys.include?(:row_sep) && !option_valid?(options[:row_sep])
-      errors << "invalid col_sep" if keys.include?(:col_sep) && !option_valid?(options[:col_sep])
-      errors << "invalid quote_char" if keys.include?(:quote_char) && !option_valid?(options[:quote_char])
-      raise SmarterCSV::ValidationError, errors.inspect if errors.any?
-    end
-
-    def option_valid?(str)
-      return true if str.is_a?(Symbol) && str == :auto
-      return true if str.is_a?(String) && !str.empty?
-
-      false
     end
   end
 end
